Rework model loading, split UI into a separate file
Large refactor part, may be broken
